CTCP-3906: Moved connector specs to it package.
