+++ it/test/connectors/ReferenceDataConnectorSpec.scala
@@ -16,31 +16,20 @@
 
 package connectors
 
-import base.{AppWithDefaultMockFixtures, SpecBase}
 import cats.data.NonEmptySet
 import com.github.tomakehurst.wiremock.client.WireMock._
 import connectors.ReferenceDataConnector.NoReferenceDataFoundException
-import helper.WireMockServerHandler
+import helpers.ItSpecBase
 import models.reference._
 import org.scalacheck.Gen
-import org.scalatest.{Assertion, BeforeAndAfterEach}
+import org.scalatest.Assertion
 import org.scalatestplus.scalacheck.ScalaCheckPropertyChecks
 import play.api.inject.guice.GuiceApplicationBuilder
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
 
-class ReferenceDataConnectorSpec
-    extends SpecBase
-    with AppWithDefaultMockFixtures
-    with WireMockServerHandler
-    with ScalaCheckPropertyChecks
-    with BeforeAndAfterEach {
-
-  override def beforeEach(): Unit = {
-    server.resetAll()
-    super.beforeEach()
-  }
+class ReferenceDataConnectorSpec extends ItSpecBase with ScalaCheckPropertyChecks {
 
   private val baseUrl = "customs-reference-data/test-only"
