+++ test/resources/application.conf
@@ -12,8 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-include "application.conf"
-
 play.filters.disabled += play.filters.csrf.CSRFFilter
 play.filters.disabled += play.filters.csp.CSPFilter
